Replace mailto with contact form
Better UX to have a message sent directly from the site than opening the
user's default mail app (which may not even be configured if they use
a browser client). Also provides some structure to the message so it's
easier to nudge users for certain data and easier to parse that data.